Refactor and include tile map template maker.
