+++ dev_tools_temp/MapMaker.java
@@ -1,5 +1,8 @@
+package dev_tools_temp;
 /**
  * MapMaker
+ * Makes a tile map JS 2D array template to the console.
+ * @author Caleb Krauter
  */
 public class MapMaker {
