+++ index.html
@@ -3,8 +3,6 @@
 
 <head>
     <title>The Good, The Bad, and Chad!</title>
-    <script type="text/javascript" src="./js/scripts/tile_maps/playground.js"></script>
-
     <!-- Scripts containing info/functions for classes. -->
     <script type="text/javascript" src="./js/scripts/util.js"></script>
     <!-- Scripts containing dialog-building methods. -->
@@ -22,6 +20,7 @@
     <script type="text/javascript" src="./js/scripts/dimension/village_dimension.js"></script>
     <script type="text/javascript" src="./js/scripts/dimension/lava_dimension.js"></script>
 
+    <!-- Tile maps here. -->
     <script type="text/javascript" src="./js/scripts/temp/bird_script.js"></script>
 
 
@@ -36,6 +35,8 @@
     <script type="text/javascript" src="./js/classes/timer.js"></script>
     <script type="text/javascript" src="./js/classes/vector.js"></script>
     <script type="text/javascript" src="./js/classes/zone.js"></script>
+
+    <script type="text/javascript" src="./js/scripts/tile_maps/caleb_zone_playground.js"></script>
 
     <!-- Entities: -->
     <!-- Background: -->
